Coordinator part 1
