+++ WMFComponents/Sources/WMFComponents/Components/Editors/Source Editor/WMFSourceEditorViewController.swift
@@ -283,14 +283,6 @@ public class WMFSourceEditorViewController: WMFComponentViewController {
     public func removeFocus() {
         textView.resignFirstResponder()
     }
-    
-    public func tappedTempAccounts() {
-        
-    }
-    
-    public func tappedIPAccounts() {
-        
-    }
 }
 
 // MARK: - Private
